+++ src/main/java/org/opensearch/knn/jni/SimdFp16.java
@@ -21,10 +21,10 @@ import static org.opensearch.knn.jni.PlatformUtils.isSIMDAVX512SPRSupportedBySys
  * Service to interact with SIMD jni layer. Class dependencies should be minimal
  * <p>
  * In order to compile C++ header file, run:
- * javac -h jni/include src/main/java/org/opensearch/knn/jni/SIMDNativeLibraryLoader.java
+ * javac -h jni/include src/main/java/org/opensearch/knn/jni/SimdFp16.java
  *      src/main/java/org/opensearch/knn/common/KNNConstants.java
  */
-public class SIMDNativeLibraryLoader {
+public class SimdFp16 {
 
     // Cached value of SIMD support
     private static final boolean SIMD_SUPPORTED;
@@ -32,9 +32,6 @@ public class SIMDNativeLibraryLoader {
     static {
         AccessController.doPrivileged((PrivilegedAction<Void>) () -> {
             try {
-                // Even if the underlying system supports AVX512 and AVX2, users can override and disable it by setting
-                // 'knn.avx2.disabled', 'knn.avx512.disabled', or 'knn.avx512_spr.disabled' to true in the opensearch.yml
-                // configuration
                 if (!isAVX512SPRDisabled() && isSIMDAVX512SPRSupportedBySystem()) {
                     System.loadLibrary(KNNConstants.SIMD_AVX512_SPR_JNI_LIBRARY_NAME);
                 } else if (!isAVX512Disabled() && isSIMDAVX512SupportedBySystem()) {
@@ -68,4 +65,26 @@ public class SIMDNativeLibraryLoader {
      * @return true if SIMD is supported and enabled, false otherwise
      */
     private static native boolean isSIMDSupportedNative();
+
+    /**
+     * Converts an array of float values to half-precision (fp16) bytes using native code.
+     *
+     * @param input float array containing float32 values
+     * @param output byte array to fill with the converted half-float values (2 bytes per value)
+     * @param count number of float values to convert
+     * @return true if native decoding succeeded (including alignment), false if fallback is required
+     */
+    public static native boolean encodeFp32ToFp16(float[] input, byte[] output, int count);
+
+    /**
+     * Converts a byte array containing float16 (half-precision) values into a float array of float32 values
+     * using native code. Uses SIMD acceleration if available, or falls back to Java if JNI returns failure.
+     *
+     * @param input byte array containing FP16-encoded values
+     * @param output float array to write the decoded FP32 values
+     * @param count number of float values to decode
+     * @param offset byte offset into the input array to begin decoding
+     * @return true if native decoding succeeded (including alignment), false if fallback is required
+     */
+    public static native boolean decodeFp16ToFp32(byte[] input, float[] output, int count, int offset);
 }